refactor: turn messages/index into a 'barrel' split the rest up
