+++ src/message/enums.ts
@@ -1,0 +1,17 @@
+export enum MessageType {
+  Connected = 'CONNECTED',
+  Disconnected = 'DISCONNECTED',
+  Entered = 'ENTERED',
+  Left = 'LEFT',
+  MovedRoom = 'MOVED',
+  SameRoom = 'SAME',
+  Chat = 'CHAT',
+  Whisper = 'WHISPER',
+  Shout = 'SHOUT',
+  Emote = 'EMOTE',
+  Dance = 'DANCE',
+  Mod = 'MOD',
+  Error = 'ERROR',
+  Command = 'COMMAND',
+  Caption = 'CAPTION'
+}
